Fix cycled imports Move StorageType to types.py
