+++ src/storages/types.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class StorageType(Enum):
+    pickle = "pickle"
+    json = "json"
+    redis = "redis"
